pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rqd-ai-lab"
version = "0.1.0"
description = "Automated Rock Quality Designation from drill core photographs using computer vision."
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "rqd-ai-lab contributors" }]
dependencies = [
    "numpy>=1.26",
    "opencv-python-headless>=4.9",
    "pillow>=10.3",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "torch>=2.3",
    "torchvision>=0.18",
    "ultralytics>=8.2",
    "mlflow>=2.14",
    "click>=8.1",
    "rich>=13.7",
    "scikit-image>=0.23",
    "scipy>=1.13",
    "pandas>=2.2",
    "tqdm>=4.66",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
    "types-tqdm",
    "interrogate>=1.7",
]
notebooks = [
    "jupyter>=1.0",
    "ipykernel>=6.29",
    "matplotlib>=3.9",
    "seaborn>=0.13",
]

[project.scripts]
rqd = "rqd_cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=50"

[tool.coverage.run]
source = ["src"]

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
fail-under = 80
